Don't generate sales orders or sales invoices if the order is already completed.
If an order is already completed, there are still some code paths that
will try to generate a new sales order or sales invoice. We don't really
want that to happen because the order is already paid for (so
recalculating taxes is a wasted effort), and also it
deletes old tax adjustments.