Set pre_tax_amount on line items when generating sales invoice
If the call to Avatax failed for some reason during SalesInvoice.generate, the order's line items would
be left without a pre_tax_amount. This causes issues when creating reimbursements, as they rely on the
line item's pre tax amount.

This updates the pre_tax_amount on each line item in the order before the call to Avatax is made. This ensures
pre_tax_amount will always be available.