Makes Avatax Less Possessive (#5)
The SpreeAvatax extension does not play well with others. It decorates several elements of the domain model, in the interest of ensuring the host app never ever ever uses any other tax-calculation mechanisms. Except...sometimes you might want to? For example, if you're expanding operations to a country not covered by Avalara (or one in which you'd rather not use Avalara).

This commit relaxes the iron grip Avatax has on the host app, such that it can co-exist with other tax calculators. The main improvements:

- relaxes the Spree::TaxRate validation rule to allow non-Avatax rates to be created
- relaxes the Spree::Adjustment validation rule to allow unfinalized non-Avatax tax adjustments
- allows Spree::Tax::OrderAdjuster to create adjustments unless in a Avatax-covered zone
- adds convenience avatax and avatax? scopes and predicates where useful (e.g. Spree:TaxRate)

Additionally, the spec suite is once again green (at least for Solidus 1.4). A more concerted effort to improve the universality of the spec suite is out of scope for this PR, but should be considered.