Accommodate Tax Adjustments with No Source (#6)
Although a tax adjustment without a source doesn't make intuitive sense, the robustness principle counsels us to be permissive in what we expect. `Spree::Adjustment#avatax?` doesn't accommodate this possibility, and Solidus doesn't enforce the presence of a source.

This commit proposes to tweak the behaviour of `Spree::Adjustment#avatax?` to check for the presence of a source before checking whether it's an Avatax source.